Best effort fork-safety in default RNG
OpenSSL library functions are generally not async-signal-safe.  There's
little point trying to make them async-signal-safe.  But there may be
apps which use OpenSSL, fork(), then continue using OpenSSL on the child
side of the fork without problems.  This patch tries to re-seed the RNG
on first use in the child just in case there is such an app.

We might want to also do this at a higher layer (i.e., in rand_lib.c
instead of md_rand.c).